+++ src/org/previmer/ichthyop/dataset/Roms2dDataset.java
@@ -133,7 +133,7 @@ public class Roms2dDataset extends RomsCommon {
     @Override
     public void nextStepTriggered(NextStepEvent e) throws Exception {
 
-        long time = e.getSource().getTime();
+        double time = e.getSource().getTime();
         //Logger.getAnonymousLogger().info("set fields at time " + time);
         int time_arrow = (int) Math.signum(e.getSource().get_dt());
 
@@ -259,7 +259,7 @@ public class Roms2dDataset extends RomsCommon {
 
     @Override
     void setOnFirstTime() throws Exception {
-        long t0 = getSimulationManager().getTimeManager().get_tO();
+        double t0 = getSimulationManager().getTimeManager().get_tO();
         ncIn = DatasetIO.openFile(DatasetIO.getFile(t0));
         readTimeLength();
         rank = findCurrentRank(t0);
