+++ src/org/previmer/ichthyop/dataset/Roms3dOpendapDataset.java
@@ -32,7 +32,7 @@ public class Roms3dOpendapDataset extends Roms3dCommon {
 
     @Override
     void setOnFirstTime() throws Exception {
-        long t0 = getSimulationManager().getTimeManager().get_tO();
+        double t0 = getSimulationManager().getTimeManager().get_tO();
         DatasetIO.checkInitTime(ncIn, strTime);
         rank = findCurrentRank(t0);
         time_tp1 = t0;
@@ -40,7 +40,7 @@ public class Roms3dOpendapDataset extends Roms3dCommon {
 
     @Override
     public void nextStepTriggered(NextStepEvent e) throws Exception {
-        long time = e.getSource().getTime();
+        double time = e.getSource().getTime();
         //Logger.getAnonymousLogger().info("set fields at time " + time);
         int time_arrow = (int) Math.signum(e.getSource().get_dt());
